+++ src/ConditionColor/constant.ts
@@ -1,0 +1,26 @@
+export const VALUE_TYPE = {
+  auto: {
+    label: '自动',
+    value: 'auto',
+  },
+  min: {
+    label: '最低值',
+    value: 'min',
+  },
+  max: {
+    label: '最高值',
+    value: 'max',
+  },
+  number: {
+    label: '数字',
+    value: 'number',
+  },
+  // percent: {
+  //   label: '百分比',
+  //   value: 'percent',
+  // },
+  // percentPoint: {
+  //   label: '百分点值',
+  //   value: 'percentPoint',
+  // },
+} as const;