+++ src/SelectTemplate/img/grade/Circle0.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+export default (
+  <svg
+    width="1em"
+    height="1em"
+    viewBox="0 0 24 24"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    data-icon="SheetIconsetsProgress0Colorful"
+  >
+    <path
+      d="M12 19.874a7.874 7.874 0 1 0 0-15.748 7.874 7.874 0 0 0 0 15.748Zm0 .926a8.8 8.8 0 1 1 0-17.6 8.8 8.8 0 0 1 0 17.6Z"
+      fill="#646A73"
+    ></path>
+  </svg>
+);