+++ src/SelectTemplate/img/grade/Mood50.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+export default (
+  <svg
+    width="1em"
+    height="1em"
+    viewBox="0 0 24 24"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    data-icon="SheetIconsetsNormalfaceColorful"
+  >
+    <path d="M20.8 12a8.8 8.8 0 1 1-17.6 0 8.8 8.8 0 0 1 17.6 0Z" fill="#FAD355"></path>
+    <path
+      d="M8.758 10.101c0-.383.31-.695.694-.695h.464a.695.695 0 0 1 0 1.39h-.464a.695.695 0 0 1-.694-.695Zm4.632 0c0-.383.31-.695.694-.695h.463a.695.695 0 1 1 0 1.39h-.463a.695.695 0 0 1-.695-.695Zm-5.698 5.141c0-.332.27-.602.602-.602v1.204a.602.602 0 0 1-.602-.602Zm.602.602h7.41a.603.603 0 0 0 .001-1.204h-7.41v1.204Z"
+      fill="#6F4A01"
+    ></path>
+  </svg>
+);