+++ src/SelectTemplate/img/shape/Prismatic.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+
+export default (
+  <svg
+    width="1em"
+    height="1em"
+    viewBox="0 0 24 24"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    data-icon="SheetIconsetsDiamondFilled"
+  >
+    <path
+      d="M4.577 12.258a.395.395 0 0 1 0-.516l7.171-8.144a.33.33 0 0 1 .504 0l7.171 8.144a.395.395 0 0 1 0 .516l-7.17 8.144a.33.33 0 0 1-.505 0l-7.171-8.144Z"
+      fill="currentColor"
+      stroke="currentColor"
+      strokeWidth="0.963"
+    ></path>
+  </svg>
+);